+++ odontologia/src/main/java/com/proyectointegrador/odontologia/repository/IOdontologoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface IOdontologoRepository extends JpaRepository<Odontologo, Integer> {
+public interface IOdontologoRepository extends JpaRepository<Odontologo, Long> {
