+++ odontologia/src/main/java/com/proyectointegrador/odontologia/repository/IPacienteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface IPacienteRepository extends JpaRepository<Paciente, Integer> {
+public interface IPacienteRepository extends JpaRepository<Paciente, Long> {
 }
